Move QRCode generation implementation to Repository
