+++ core/data/build.gradle.kts
@@ -29,6 +29,7 @@ kotlin {
                 implementation(libs.ktorKotlinxSerialization)
                 implementation(libs.ktorContentNegotiation)
                 implementation(libs.kermit)
+                implementation(libs.qrcodeKotlin)
             }
         }
