+++ core/data/src/commonMain/kotlin/io/github/droidkaigi/confsched/data/profilecard/DefaultProfileCardRepository.kt
@@ -7,10 +7,16 @@ import io.github.droidkaigi.confsched.compose.safeCollectAsRetainedState
 import io.github.droidkaigi.confsched.model.ProfileCard
 import io.github.droidkaigi.confsched.model.ProfileCardRepository
 import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.withContext
+import org.jetbrains.compose.resources.DrawableResource
+import org.jetbrains.compose.resources.ExperimentalResourceApi
+import org.jetbrains.compose.resources.getDrawableResourceBytes
+import org.jetbrains.compose.resources.getSystemResourceEnvironment
+import qrcode.QRCode
 
 public class DefaultProfileCardRepository(
     private val profileCardDataStore: ProfileCardDataStore,
-    override val ioDispatcher: CoroutineDispatcher,
+    private val ioDispatcher: CoroutineDispatcher,
 ) : ProfileCardRepository {
     @Composable
     override fun profileCard(): ProfileCard {
@@ -23,5 +29,19 @@ public class DefaultProfileCardRepository(
 
     override suspend fun save(profileCard: ProfileCard.Exists) {
         profileCardDataStore.save(profileCard)
+    }
+
+    @OptIn(ExperimentalResourceApi::class)
+    override suspend fun loadQrCodeImageByteArray(link: String, centerLogoRes: DrawableResource): ByteArray {
+        return withContext(ioDispatcher) {
+            val logoImage = getDrawableResourceBytes(
+                environment = getSystemResourceEnvironment(),
+                resource = centerLogoRes,
+            )
+            QRCode.ofSquares()
+                .withLogo(logoImage, 400, 400)
+                .build(link)
+                .renderToBytes()
+        }
     }
 }
