+++ core/model/src/commonMain/kotlin/io/github/droidkaigi/confsched/model/ProfileCardRepository.kt
@@ -3,13 +3,13 @@ package io.github.droidkaigi.confsched.model
 import androidx.compose.runtime.Composable
 import io.github.droidkaigi.confsched.model.compositionlocal.LocalRepositories
 import kotlinx.coroutines.CoroutineDispatcher
+import org.jetbrains.compose.resources.DrawableResource
 
 interface ProfileCardRepository {
-    val ioDispatcher: CoroutineDispatcher
-
     @Composable
     fun profileCard(): ProfileCard
     suspend fun save(profileCard: ProfileCard.Exists)
+    suspend fun loadQrCodeImageByteArray(link: String, centerLogoRes: DrawableResource): ByteArray
 }
 
 @Composable
