+++ feature/profilecard/build.gradle.kts
@@ -16,7 +16,6 @@ kotlin {
                 implementation(libs.composeNavigation)
                 implementation(compose.materialIconsExtended)
                 implementation(libs.peekabooImagePicker)
-                implementation(libs.qrcodeKotlin)
             }
         }
         androidTarget {
